Add files via upload
PROJECT DESCRIPTION: The Project is based on Operation Analytics, case-1 dataset is given with all variables of the project. Initially I will go through the given dataset and see the observation of each and every variable table and formulate the output for all perspective tasks.  

APPROACH: I will study the whole dataset for the given Operation Analytics users, Analyse the different data inside the table and note down the total number of observations for different data available and I will also execute the data for the required task which is mentioned for each task.

TECH-STACK USED: My SQL Workbench -8.0 CE
I used My SQL Workbench, it's a software for SQL database playground for testing, debugging and executing the code. The user interface is easy to understand, and we can execute the code within minimal time. 

INSIGHTS: With the help of application My SQL Workbench, I have imported the file and executed it. Initially I looked in to all the variables of the database. (job_id,actor_id,event,language,time_spent,org,ds). with respect to the target task, I have written code line to obtain all the data in tables for tasks.